pyproject.toml
[project]
name = "phoebe-pose"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "zku" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.9.0",
    "pillow>=11.1.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "ruff>=0.11.4",
    "streamlit>=1.44.1",
]

[project.scripts]
phoebe-pose = "phoebe_pose:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"

[browser]
gatherUsageStats = false
